Handle unknown report types correctly
Before if you parsed a vector of reports, _any_ unknown report type
would result in a parse error that threw the entire result vector away.
This patch simplifies things, throwing away all of the `Any`-based
shenanigans.  When parsing, we just stash the body away into a JSON
value.  Later on, when you're trying to operate on a particular kind of
report, we try to parse the body using that report type's schema.